Add SettingsDatastoreRepository in preparation for adding console inputs as an optional setting.
Refactor individual layouts in combo creation screen for each game available.

Adjust UI scaling on landscape mode from 2x to 1.5x to allow for longer combos on a smaller screen.

Add Inputs for SF6 and update Inputs for MK and Tekken to work better with the layouts.

Bugs to fix:

Cannot change orientation on AddComboScreen as you lose Combo details. Orientation will be locked until this issue is resolved.