Create AddCharacterScreen and AddCharacterViewModel
Create logic in for custom character entry which adds custom characters to the database. If the character exists, the user will see a snackbar asking if they want to update the existing entry.

Update CharacterEntry to check if a character is mutable, if so, that character can't be updated in the custom character screen.

Bugs to fix:

Cannot change orientation on AddComboScreen as you lose Combo details. Orientation will be locked until this issue is resolved.